// src/common/enums/productType.ts
export enum ProductType {
  NO_VARIABLE_NO_SERIAL = 100, //Không có biến thể không có seri
  NORMAL_SERIALS = 200, // Không có biến thể có seri
  NORMAL_VARIABLES = 300, //Có biến thể không có seri
  NORMAL_VARIABLES_SERIALS = 400, // Có biến thể trong biến thể có seri
}